requirements.txt
pycryptodome==3.15.0      
cryptography==3.4.7         
flask==2.0.1               
numpy==1.21.2               
sympy==1.9                 
matplotlib==3.4.3          
requests==2.26.0            

#To install these dependencies, simply run the following command:

pip install -r requirements.txt
